refactor: move apply_schema out of arrow_expression
Follow-up to #2558. apply_schema is consumed by the parquet read path
(arrow_utils) as well as the expression evaluator — promote it to a
top-level engine module so the location matches its scope.
