+++ kernel/src/engine/arrow_expression/mod.rs
@@ -1,7 +1,6 @@
 //! Expression handling based on arrow-rs compute kernels.
 use std::sync::Arc;
 
-use apply_schema::{apply_schema, apply_schema_to};
 use evaluate_expression::{evaluate_expression, evaluate_predicate, extract_column};
 use itertools::Itertools;
 use tracing::debug;
@@ -12,14 +11,13 @@ use crate::arrow::datatypes::{
     DataType as ArrowDataType, Field as ArrowField, Schema as ArrowSchema,
 };
 use crate::engine::arrow_data::{extract_record_batch, ArrowEngineData};
+use crate::engine::arrow_utils::apply_schema::{apply_schema, apply_schema_to};
 use crate::error::{DeltaResult, Error};
 use crate::expressions::{ArrayData, Expression, ExpressionRef, PredicateRef, Scalar};
 use crate::schema::{DataType, PrimitiveType, SchemaRef};
 use crate::utils::require;
 use crate::{EngineData, EvaluationHandler, ExpressionEvaluator, PredicateEvaluator};
 
-mod apply_schema;
-pub(crate) use apply_schema::apply_schema_to_struct;
 pub mod evaluate_expression;
 pub mod opaque;
